% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drivers}
\alias{drivers}
\title{Brazilian Licensed Drivers Data}
\format{
\subsection{\code{drivers}}{

A \code{tibble} data frame encoded to Brazilian Portuguese:
\describe{
\item{regiao}{Brazilian Geographic Region}
\item{uf}{Federative Unit - Brazilian States Acronym}
\item{sexo}{Drivers sex}
\item{faixa_etaria}{Drivers age group}
\item{ano}{Year}
\item{categoria_cnh}{Drivers license category}
\item{condutores}{Amount of drivers}
}
}
}
\source{
https://www.gov.br/transportes/pt-br/assuntos/transito/conteudo-Senatran/estatisticas-quantidade-de-habilitados-denatran
}
\usage{
drivers
}
\description{
A dataset of licensed drivers data scraped from the
Brazilian Ministry of Transport
}
\keyword{datasets}
